requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
firebase-admin>=6.2.0
requests>=2.28.0
Pillow>=9.5.0
python-dotenv>=1.0.0
 